# src/main/java/com/zufar/icedlatte/review/endpoint/ProductReviewEndpoint.java
package com.zufar.icedlatte.review.endpoint;

import com.zufar.icedlatte.openapi.dto.ProductReviewRateDto;
import com.zufar.icedlatte.openapi.dto.ProductReviewRatingStats;
import com.zufar.icedlatte.openapi.dto.ProductReviewRequest;
import com.zufar.icedlatte.openapi.dto.ProductReviewResponse;
import com.zufar.icedlatte.openapi.dto.ProductReviewsAndRatingsWithPagination;
import com.zufar.icedlatte.review.api.ProductReviewRateProvider;
import com.zufar.icedlatte.review.api.ProductReviewRateUpdater;
import com.zufar.icedlatte.review.api.ProductReviewsProvider;
import com.zufar.icedlatte.review.api.ProductReviewCreator;
import com.zufar.icedlatte.review.api.ProductReviewDeleter;
import com.zufar.icedlatte.review.api.ProductReviewsStatisticsProvider;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;
import java.util.UUID;

@Slf4j
@RestController
@RequiredArgsConstructor
@Validated
@RequestMapping(value = ProductReviewEndpoint.PRODUCT_REVIEW_URL)
public class ProductReviewEndpoint implements com.zufar.icedlatte.openapi.product.review.api.ProductReviewApi {

    public static final String PRODUCT_REVIEW_URL = "/api/v1/products/";

    private final ProductReviewCreator productReviewCreator;
    private final ProductReviewDeleter productReviewDeleter;
    private final ProductReviewsProvider productReviewsProvider;
    private final ProductReviewsStatisticsProvider productReviewsStatisticsProvider;
    private final ProductReviewRateUpdater productReviewRateUpdater;
    private final ProductReviewRateProvider productReviewRateProvider;

    @Override
    @PostMapping(value = "/{productId}/reviews")
    public ResponseEntity<ProductReviewResponse> addNewProductReview(@PathVariable final UUID productId,
                                                                     final ProductReviewRequest productReviewRequest) {
        log.info("Received the request to add a new product review for the product with the productId = '{}'", productId);
        var review = productReviewCreator.create(productId, productReviewRequest);
        log.info("New product review was added with productReviewId = '{}' for the product with the productId = '{}'", review.getProductReviewId(), productId);
        return ResponseEntity.ok().body(review);
    }

    @Override
    @DeleteMapping(value = "/{productId}/reviews/{productReviewId}")
    public ResponseEntity<Void> deleteProductReview(@PathVariable final UUID productId,
                                                    @PathVariable final UUID productReviewId) {
        log.info("Received request to delete product review with productReviewId = '{}', productId = '{}'", productReviewId, productId);
        productReviewDeleter.delete(productId, productReviewId);
        log.info("Product review with productReviewId = '{}' was deleted", productReviewId);
        return new ResponseEntity<>(HttpStatus.OK);
    }

    @Override
    @GetMapping(value = "/{productId}/reviews")
    public ResponseEntity<ProductReviewsAndRatingsWithPagination> getProductReviewsAndRatings(@PathVariable final UUID productId,
                                                                                              @RequestParam(name = "page", defaultValue = "0") final Integer page,
                                                                                              @RequestParam(name = "size", defaultValue = "10") final Integer size,
                                                                                              @RequestParam(name = "sort_attribute", defaultValue = "createdAt") final String sortAttribute,
                                                                                              @RequestParam(name = "sort_direction", defaultValue = "desc") final String sortDirection) {
        log.info("Received the request to get reviews and ratings for the product with the productId = '{}' and with the next pagination and sorting attributes: page - {}, size - {}, sort_attribute - {}, sort_direction - {}",
                productId, page, size, sortAttribute, sortDirection);
        ProductReviewsAndRatingsWithPagination reviewsPaginationDto = productReviewsProvider.getProductReviews(productId, page, size, sortAttribute, sortDirection);
        log.info("Product reviews and ratings were retrieved successfully for the product with the productId = '{}'", productId);
        return ResponseEntity.ok().body(reviewsPaginationDto);
    }

    @Override
    @GetMapping(value = "/{productId}/review")
    public ResponseEntity<ProductReviewResponse> getProductReview(@PathVariable final UUID productId) {
        log.info("Received the request to get product review and rating for the product with the productId = '{}'", productId);
        ProductReviewResponse result = productReviewsProvider.getProductReviewForUser(productId);
        log.info("Product review and rating were retrieved successfully for the product with the productId = '{}'", productId);
        return ResponseEntity.ok().body(result);
    }

    @Override
    @GetMapping("/{productId}/reviews/statistics")
    public ResponseEntity<ProductReviewRatingStats> getRatingAndReviewStat(@PathVariable final UUID productId) {
        log.info("Received the request to get the statistics of product's review and rating for the product with the productId = '{}'", productId);
        final ProductReviewRatingStats stats = productReviewsStatisticsProvider.getRatingAndReviewStat(productId);
        log.info("Statistics for product's review and rating for the product with the productId = '{}' was retrieved successfully", productId);
        return ResponseEntity.ok().body(stats);
    }

    @Override
    @PostMapping(value = "/{productId}/reviews/rate")
    public ResponseEntity<ProductReviewRateDto> rateProductReview(@PathVariable final UUID productId,
                                                                  final ProductReviewRateDto request) {
        log.info("Received the request to rate a product review for the product with the productId = '{}'", productId);
        var reviewRate = productReviewRateUpdater.updateReviewRate(productId, request);
        log.info("Product review with id {} was successfully {}", request.getProductReviewId(), request.getIsLike() ? "liked" : "disliked");
        return ResponseEntity.ok().body(reviewRate);
    }

    @Override
    @GetMapping(value = "/{productId}/reviews/rate")
    public ResponseEntity<List<ProductReviewRateDto>> getProductReviewRates(@PathVariable final UUID productId) {
        log.info("Received the request to fetch product review rates for the product with the productId = '{}'", productId);
        var reviewRates = productReviewRateProvider.getProductReviewRates(productId);
        log.info("Product review rates were successfully retrieved");
        return ResponseEntity.ok().body(reviewRates);
    }
}